+++ projects/core/components/scrollbar/scrollbar.directive.ts
@@ -1,112 +1,37 @@
 import {Directive, inject, input} from '@angular/core';
 import {takeUntilDestroyed} from '@angular/core/rxjs-interop';
-import {WA_ANIMATION_FRAME} from '@ng-web-apis/common';
-import {
-    tuiScrollFrom,
-    tuiZonefree,
-    tuiZonefreeScheduler,
-} from '@taiga-ui/cdk/observables';
 import {tuiInjectElement} from '@taiga-ui/cdk/utils/dom';
-import {merge, throttleTime} from 'rxjs';
 
 import {TUI_SCROLL_REF} from './scroll-ref.directive';
 import {TuiScrollbarService} from './scrollbar.service';
-
-const MIN_WIDTH = 24;
-
-interface ComputedDimension {
-    scrollTop: number;
-    scrollHeight: number;
-    clientHeight: number;
-    scrollLeft: number;
-    scrollWidth: number;
-    clientWidth: number;
-}
+import {TuiScrollbarPosition} from './scrollbar-position.directive';
 
 @Directive({
     selector: '[tuiScrollbar]',
     providers: [TuiScrollbarService],
+    hostDirectives: [{directive: TuiScrollbarPosition, inputs: ['tuiScrollbar']}],
 })
 export class TuiScrollbarDirective {
-    private readonly el = inject(TUI_SCROLL_REF).nativeElement;
+    private readonly scrollRef = inject(TUI_SCROLL_REF);
     private readonly style = tuiInjectElement().style;
 
     protected readonly scrollSub = inject(TuiScrollbarService)
         .pipe(takeUntilDestroyed())
         .subscribe(([top, left]) => {
-            this.el.style.scrollBehavior = 'auto';
+            this.scrollRef.nativeElement.style.scrollBehavior = 'auto';
 
             if (this.tuiScrollbar() === 'horizontal') {
-                this.el.scrollLeft = left;
+                this.scrollRef.nativeElement.scrollLeft = left;
             } else {
-                this.el.scrollTop = top;
+                this.scrollRef.nativeElement.scrollTop = top;
             }
 
-            this.el.style.scrollBehavior = '';
+            this.scrollRef.nativeElement.style.scrollBehavior = '';
         });
 
-    protected readonly styleSub = merge(
-        inject(WA_ANIMATION_FRAME).pipe(throttleTime(100, tuiZonefreeScheduler())),
-        tuiScrollFrom(this.el),
-    )
-        .pipe(tuiZonefree(), takeUntilDestroyed())
-        .subscribe(() => {
-            const dimension: ComputedDimension = {
-                scrollTop: this.el.scrollTop,
-                scrollHeight: this.el.scrollHeight,
-                clientHeight: this.el.clientHeight,
-                scrollLeft: this.el.scrollLeft,
-                scrollWidth: this.el.scrollWidth,
-                clientWidth: this.el.clientWidth,
-            };
-
-            const thumb = `${this.getThumb(dimension) * 100}%`;
-            const view = `${this.getView(dimension) * 100}%`;
-
-            if (this.tuiScrollbar() === 'vertical') {
-                this.style.top = thumb;
-                this.style.height = view;
-            } else {
-                this.style.left = thumb;
-                this.style.insetInlineStart = thumb;
-                this.style.width = view;
-            }
-        });
+    protected readonly styleSub = inject(TuiScrollbarPosition)
+        .pipe(takeUntilDestroyed())
+        .subscribe((position) => Object.assign(this.style, position));
 
     public readonly tuiScrollbar = input<'horizontal' | 'vertical'>('vertical');
-
-    private getScrolled(dimension: ComputedDimension): number {
-        return this.tuiScrollbar() === 'vertical'
-            ? dimension.scrollTop / (dimension.scrollHeight - dimension.clientHeight)
-            : dimension.scrollLeft / (dimension.scrollWidth - dimension.clientWidth);
-    }
-
-    private getCompensation(dimension: ComputedDimension): number {
-        if (
-            ((dimension.clientHeight * dimension.clientHeight) / dimension.scrollHeight >
-                MIN_WIDTH &&
-                this.tuiScrollbar() === 'vertical') ||
-            ((dimension.clientWidth * dimension.clientWidth) / dimension.scrollWidth >
-                MIN_WIDTH &&
-                this.tuiScrollbar() === 'horizontal')
-        ) {
-            return 0;
-        }
-
-        return this.tuiScrollbar() === 'vertical'
-            ? MIN_WIDTH / dimension.clientHeight
-            : MIN_WIDTH / dimension.clientWidth;
-    }
-
-    private getThumb(dimension: ComputedDimension): number {
-        const compensation = this.getCompensation(dimension) || this.getView(dimension);
-
-        return Math.abs(this.getScrolled(dimension) * (1 - compensation));
-    }
-
-    private getView(dimension: ComputedDimension): number {
-        return this.tuiScrollbar() === 'vertical'
-            ? Math.ceil((dimension.clientHeight / dimension.scrollHeight) * 100) / 100
-            : Math.ceil((dimension.clientWidth / dimension.scrollWidth) * 100) / 100;
-    }
 }
